requirements.txt
scrapy
pyjanitor
pip-install-test
html5lib
nameparser
